exam_generator/requirements.txt
pandas
openpyxl
python-docx
tkinter